If a user enters a bad URL, preserve it on the error page
Previously typing in a URL which was rejected because:

* it wasn't on flickr.com, e.g. example.net, or
* it wasn't a photo page, e.g. flickr.com/help

would be rejected and clear the input field.  This is annoying, because
the user can't fix their input or even see what they typed!

Passing the `flickr_url` variable to the template ensures the problem
URL is still shown on the page when they get the error.