Add the categories after the initial upload step
This is technically less efficient (we could include categories in the
first call to the Upload API) but it's required to make the page actually
look right immediately after upload, and before any other edits.

I have tried using the Purge API command for this, but it didn't have the
desired effect.