Simplify the way we bucket the matches
It seems like the Wikimedia API usually returns matches based on
the canonical label if it's at the start of the label (or at least the
start of a word).

e.g. a search for "cym" turns up "Cymraeg", but a search for "ymr" is
empty -- it can't look inside words.

Given that the majority of language labels are single words, this means
all the matches will be at the start.  Distinguishing between matches
that occur after the start feels unnecessary.